requirements.txt
numpy
matplotlib
scipy
torch
torchvision
transformers
torchmetrics
tqdm
requests
colorama
comet_ml
kaggle
python-dotenv
ray[tune]
PyYaml
seaborn
timm
einops
openxlab
open_clip_torch
umap-learn[plot]
pykeops
torchdr
datamapplot